refactor: remove trace_decoder crate
