--- trace_decoder/Cargo.toml
@@ -1,81 +1,0 @@
-[package]
-name = "trace_decoder"
-description = "Ethereum node witness -> Prover input"
-authors = ["Polygon Zero"]
-version = "0.6.0"
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-homepage.workspace = true
-keywords.workspace = true
-
-[dependencies]
-alloy = { workspace = true }
-alloy-compat = "0.1.0"
-anyhow = { workspace = true }
-bitflags = { workspace = true }
-bitvec = { workspace = true }
-bytes = { workspace = true }
-ciborium = { workspace = true }
-ciborium-io = { workspace = true }
-copyvec = "0.2.0"
-either = { workspace = true }
-enum-as-inner = { workspace = true }
-ethereum-types = { workspace = true }
-hex = { workspace = true }
-hex-literal = { workspace = true }
-itertools = { workspace = true }
-keccak-hash = { workspace = true }
-log = { workspace = true }
-nunny = { workspace = true, features = ["serde"] }
-plonky2 = { workspace = true }
-rlp = { workspace = true }
-serde = { workspace = true }
-stackstack = "0.3.0"
-strum = { version = "0.26.3", features = ["derive"] }
-thiserror = { workspace = true }
-u4 = { workspace = true }
-winnow = { workspace = true }
-
-# Local dependencies
-evm_arithmetization = { workspace = true }
-mpt_trie = { workspace = true }
-smt_trie = { workspace = true }
-zk_evm_common = { workspace = true }
-
-[dev-dependencies]
-alloy = { workspace = true }
-alloy-compat = "0.1.0"
-assert2 = "0.3.15"
-camino = "1.1.9"
-clap = { workspace = true }
-criterion = { workspace = true }
-glob = "0.3.1"
-libtest-mimic = "0.7.3"
-plonky2_maybe_rayon = { workspace = true }
-pretty_assertions = "1.4.0"
-zero = { workspace = true }
-pretty_env_logger = { workspace = true }
-serde_json = { workspace = true }
-serde_path_to_error = { workspace = true }
-
-[features]
-default = ["eth_mainnet"]
-eth_mainnet = ["evm_arithmetization/eth_mainnet", "zero/eth_mainnet"]
-cdk_erigon = ["evm_arithmetization/cdk_erigon", "zero/cdk_erigon"]
-polygon_pos = ["evm_arithmetization/polygon_pos", "zero/polygon_pos"]
-
-[[bench]]
-name = "block_processing"
-harness = false
-required-features = ["eth_mainnet"]
-
-[[test]]
-name = "consistent-with-header"
-harness = false
-required-features = ["eth_mainnet"]
-
-[[test]]
-name = "simulate-execution"
-harness = false
-required-features = ["eth_mainnet"]
