+++ zero/src/bin/rpc.rs
@@ -8,11 +8,11 @@ use alloy::transports::Transport;
 use anyhow::anyhow;
 use clap::{Args, Parser, Subcommand, ValueHint};
 use futures::StreamExt;
-use trace_decoder::observer::DummyObserver;
 use tracing_subscriber::{prelude::*, EnvFilter};
 use url::Url;
 use zero::block_interval::BlockInterval;
 use zero::block_interval::BlockIntervalStream;
+use zero::intra_block_tries::DummyObserver;
 use zero::prover::BlockProverInput;
 use zero::provider::CachedProvider;
 use zero::rpc;
@@ -167,7 +167,7 @@ impl Cli {
                                 block_number
                             ))?;
 
-                        let generation_inputs = trace_decoder::entrypoint(
+                        let generation_inputs = zero::intra_block_tries::entrypoint(
                             block_prover_input.block_trace,
                             block_prover_input.other_data,
                             batch_size,
