+++ zero/src/wire_tries/cases/hermez_cdk_erigon.json
